user + address link + db + refactoring

+++ user-service/src/main/java/com/vlados/webshop/userservice/service/impl/UserServiceImpl.java
@@ -1,11 +1,14 @@
 package com.vlados.webshop.userservice.service.impl;
 
+import com.vlados.webshop.userservice.dao.AddressDao;
 import com.vlados.webshop.userservice.dao.UserDao;
+import com.vlados.webshop.userservice.dto.address.AddressDto;
 import com.vlados.webshop.userservice.dto.user.NewUserDto;
 import com.vlados.webshop.userservice.dto.user.ResponseUserDto;
 import com.vlados.webshop.userservice.dto.user.UpdatedUserDto;
 import com.vlados.webshop.userservice.service.UserService;
 import com.vlados.webshop.userservice.util.ResourceUtil;
+import com.vlados.webshop.userservice.util.mapper.AddressMapper;
 import com.vlados.webshop.userservice.util.mapper.UserMapper;
 import org.springframework.stereotype.Service;
 
@@ -16,9 +19,11 @@ import java.util.Optional;
 @Service
 public class UserServiceImpl implements UserService {
     private final UserDao userDao;
+    private final AddressDao addressDao;
 
-    public UserServiceImpl(UserDao userDao) {
+    public UserServiceImpl(UserDao userDao, AddressDao addressDao) {
         this.userDao = userDao;
+        this.addressDao = addressDao;
     }
 
     @Override
@@ -57,17 +62,33 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void delete(final long id) {
-        userDao.delete(id);
+    public void delete(final long id) throws NoSuchElementException {
+        if (exists(id)) {
+            userDao.delete(id);
+        } else {
+            throw new NoSuchElementException(ResourceUtil.getMessage("db.user.id").formatted(id));
+        }
     }
 
     @Override
-    public void update(final long id, final UpdatedUserDto dto) {
-        userDao.update(id, dto);
+    public void update(final long id, final UpdatedUserDto dto) throws NoSuchElementException {
+        if (exists(id)) {
+            userDao.update(id, dto);
+            AddressDto newAddressDto = dto.address();
+            if (addressChanged(id, newAddressDto)) {
+                userDao.update(id, newAddressDto);
+            }
+        } else {
+            throw new NoSuchElementException(ResourceUtil.getMessage("db.user.id").formatted(id));
+        }
     }
 
     @Override
     public boolean exists(final long id) {
         return userDao.exists(id);
+    }
+
+    private boolean addressChanged(long id, AddressDto dto) {
+        return !dto.equals(AddressMapper.map(addressDao.getForUser(id).get()));
     }
 }
